fix: 값 변경 안되는 버그 수정 완료
스코어 매니저 수정

+++ Assets/08_Scripts/UI/Score/IScorable.cs
@@ -18,7 +18,7 @@ public interface IScorable
 {
     void SetScore(ScoreType scoreType, int score);
 
-    float GetScore(ScoreType scoreType);
+    int GetScore(ScoreType scoreType);
 
     bool IsScorePerfect(ScoreType scoreType);
 }
